requirements.txt
fastapi~=0.73.0
uvicorn~=0.17.5
pydantic~=1.9.0
typing~=3.7.4.3
pyhumps~=3.5.3
sqlalchemy~=1.4.34
pymysql~=1.0.2